Implement ECS: Integrate DescribesIntances and DescribeNetworkInterfaces into one class AwsEc2Api
